Move the merging algorithm to comparison.js
