+++ src/actions/receive-profile.js
@@ -7,18 +7,11 @@ import { oneLine } from 'common-tags';
 import {
   processProfile,
   unserializeProfileOfArbitraryFormat,
-  adjustSampleTimestamps,
-  adjustMarkerTimestamps,
 } from '../profile-logic/process-profile';
 import { SymbolStore } from '../profile-logic/symbol-store';
-import { getEmptyProfile } from '../profile-logic/data-structures';
-import {
-  filterThreadSamplesToRange,
-  getTimeRangeForThread,
-  getTimeRangeIncludingAllThreads,
-} from '../profile-logic/profile-data';
 import { symbolicateProfile } from '../profile-logic/symbolication';
 import * as MozillaSymbolicationAPI from '../profile-logic/mozilla-symbolication-api';
+import { mergeProfiles } from '../profile-logic/comparison';
 import { decompress } from '../utils/gz';
 import { TemporaryError } from '../utils/errors';
 import JSZip from 'jszip';
@@ -43,7 +36,6 @@ import {
   initializeHiddenGlobalTracks,
   getVisibleThreads,
 } from '../profile-logic/tracks';
-import { mergeCategories, adjustCategories } from '../profile-logic/comparison';
 
 import type {
   FunctionsUpdatePerThread,
@@ -57,7 +49,6 @@ import type {
   Profile,
   ThreadIndex,
   IndexIntoFuncTable,
-  IndexIntoCategoryList,
 } from '../types/profile';
 
 /**
@@ -872,100 +863,12 @@ export function retrieveProfilesToCompare(
       // pushing them to a brand new profile. This resulting profile will keep
       // only the 2 selected threads from the 2 profiles.
       const profiles = await Promise.all(promises);
-      const resultProfile = getEmptyProfile();
 
-      // First let's merge categories. We'll use the resulting maps when
-      // handling the thread data later.
       const {
-        categories: newCategories,
-        translationMaps: translationMapsForCategories,
-      } = mergeCategories(profiles.map(profile => profile.meta.categories));
-      resultProfile.meta.categories = newCategories;
-
-      // Then we loop over all profiles and do the necessary changes according
-      // to the states we computed earlier.
-      const transformStacks = {};
-      const implementationFilters = [];
-
-      for (let i = 0; i < profileStates.length; i++) {
-        const { profileSpecific } = profileStates[i];
-        const selectedThreadIndex = profileSpecific.selectedThread;
-        if (selectedThreadIndex === null) {
-          throw new Error(`No thread has been selected in profile ${i}`);
-        }
-        const profile = profiles[i];
-        let thread = profile.threads[selectedThreadIndex];
-        transformStacks[i] = profileSpecific.transforms[selectedThreadIndex];
-        implementationFilters.push(profileSpecific.implementation);
-
-        // We adjust the categories using the maps computed above.
-        // Here we're cheating a bit with flow here because we know that we
-        // can't have null values in this table.
-        thread.stackTable.category = ((adjustCategories(
-          thread.stackTable.category,
-          translationMapsForCategories[i]
-        ): any): IndexIntoCategoryList[]);
-        thread.frameTable.category = adjustCategories(
-          thread.frameTable.category,
-          translationMapsForCategories[i]
-        );
-
-        // We filter the profile using the range from the state for this profile.
-        const zeroAt = getTimeRangeIncludingAllThreads(profile).start;
-        const committedRange =
-          profileSpecific.committedRanges &&
-          profileSpecific.committedRanges.pop();
-        thread = committedRange
-          ? filterThreadSamplesToRange(
-              thread,
-              committedRange.start + zeroAt,
-              committedRange.end + zeroAt
-            )
-          : thread;
-
-        // We're reseting the thread's PID to make sure we don't have any collision.
-        thread.pid = `${thread.pid} from profile ${i + 1}`;
-        thread.processName = `Profile ${i + 1}: ${thread.processName ||
-          thread.name}`;
-
-        // We adjust the various times so that the 2 profiles are aligned at the
-        // start and the data is consistent.
-        const startTimeAdjustment = -thread.samples.time[0];
-        thread.samples = adjustSampleTimestamps(
-          thread.samples,
-          startTimeAdjustment
-        );
-        thread.markers = adjustMarkerTimestamps(
-          thread.markers,
-          startTimeAdjustment
-        );
-        thread.registerTime += startTimeAdjustment;
-        thread.processStartupTime += startTimeAdjustment;
-        if (thread.processShutdownTime !== null) {
-          thread.processShutdownTime += startTimeAdjustment;
-        }
-        if (thread.unregisterTime !== null) {
-          thread.unregisterTime += startTimeAdjustment;
-        }
-
-        // The loaded profiles will often have different lengths. We align the
-        // start times in the block above, so this means the end times will be
-        // different.
-        // By setting `unregisterTime` here, the empty thread indicators will be
-        // drawn, which will help the users visualizing the different lengths of
-        // the loaded profiles.
-        if (
-          thread.processShutdownTime === null &&
-          thread.unregisterTime === null
-        ) {
-          thread.unregisterTime = getTimeRangeForThread(
-            thread,
-            profile.meta.interval
-          ).end;
-        }
-
-        resultProfile.threads.push(thread);
-      }
+        profile: resultProfile,
+        implementationFilters,
+        transformStacks,
+      } = mergeProfiles(profiles, profileStates);
 
       // We define an implementationFilter if both profiles agree with the value.
       let implementationFilter;
